refactor: move price impact logic into new price-impact module
