+++ app/(app)/pools/[chain]/[variant]/[id]/add-liquidity/[[...txHash]]/layout.tsx
@@ -4,7 +4,7 @@ import { isNotSupported } from '@/lib/modules/pool/pool.helpers'
 import { usePool } from '@/lib/modules/pool/usePool'
 import { RelayerSignatureProvider } from '@/lib/modules/relayer/useRelayerSignature'
 import { TokenInputsValidationProvider } from '@/lib/modules/tokens/useTokenInputsValidation'
-import { PriceImpactProvider } from '@/lib/shared/hooks/usePriceImpact'
+import { PriceImpactProvider } from '@/lib/modules/price-impact/usePriceImpact'
 import { Alert } from '@chakra-ui/react'
 import { AddLiquidityProvider } from '../../../../../../../../lib/modules/pool/actions/add-liquidity/useAddLiquidity'
 import { TransactionStateProvider } from '@/lib/modules/transactions/transaction-steps/TransactionStateProvider'
