+++ lib/modules/pool/actions/add-liquidity/useAddLiquidity.tsx
@@ -28,7 +28,7 @@ import { useAddLiquiditySteps } from './useAddLiquiditySteps'
 import { useTransactionSteps } from '@/lib/modules/transactions/transaction-steps/useTransactionSteps'
 import { useTotalUsdValue } from '@/lib/modules/tokens/useTotalUsdValue'
 import { HumanTokenAmountWithAddress } from '@/lib/modules/tokens/token.types'
-import { isUnhandledAddPriceImpactError } from '@/lib/shared/utils/price-impact'
+import { isUnhandledAddPriceImpactError } from '@/lib/modules/price-impact/priceImpact.helpers'
 
 export type UseAddLiquidityResponse = ReturnType<typeof _useAddLiquidity>
 export const AddLiquidityContext = createContext<UseAddLiquidityResponse | null>(null)
