+++ lib/modules/price-impact/PriceImpactAcceptModal.tsx
