+++ lib/modules/swap/SwapDetails.tsx
@@ -6,7 +6,7 @@ import { HStack, VStack, Text, Tooltip, Box } from '@chakra-ui/react'
 import { useSwap } from './useSwap'
 import { GqlSorSwapType } from '@/lib/shared/services/api/generated/graphql'
 import { useUserSettings } from '../user/settings/useUserSettings'
-import { usePriceImpact } from '@/lib/shared/hooks/usePriceImpact'
+import { usePriceImpact } from '@/lib/modules/price-impact/usePriceImpact'
 import { SdkSimulateSwapResponse } from './swap.types'
 import { DefaultSwapHandler } from './handlers/DefaultSwap.handler'
 import { useTokens } from '../tokens/useTokens'
