+++ lib/modules/swap/SwapForm.tsx
@@ -23,8 +23,8 @@ import { isSameAddress } from '@/lib/shared/utils/addresses'
 import { Address } from 'viem'
 import { SwapPreviewModal } from './SwapPreviewModal'
 import { TransactionSettings } from '../user/settings/TransactionSettings'
-import { PriceImpactAccordion } from '../../shared/components/accordion/PriceImpactAccordion'
-import { PriceImpactProvider } from '@/lib/shared/hooks/usePriceImpact'
+import { PriceImpactAccordion } from '../price-impact/PriceImpactAccordion'
+import { PriceImpactProvider } from '@/lib/modules/price-impact/usePriceImpact'
 import { ChainSelect } from '../chains/ChainSelect'
 import { CheckCircle, Link, Repeat } from 'react-feather'
 import { SwapRate } from './SwapRate'
